pyproject.toml
[project]
name = "cc-benchmark"
version = "1.0.0"
description = "SWE-Bench Plugin Efficiency Benchmark Tool - measures plugin/tool efficiency across success rate, execution time, tool calls, and token usage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "datasets>=2.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
swe-bench-harness = "swe_bench_harness.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["swe_bench_harness"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
